[SYCL] Add regression test for std::vector in device code
When std containers are used, FE emits several Intel-specific intrinsics
which are later used by optimizer, but those intrinsics are not
supported in SPIR-V and therefore we weren't able to compile some code
samples until we enabled early optimizatoins in SYCL by default.

This commit adds a regression test to ensure that we are still able to
compile such code example.

Change-Id: Ie7b78fa7f0f88f5780d05071cddd473e8299b145
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/302307
Tested-by: nbtester <nbtester@intel.com>
Reviewed-by: Sachkov, Alexey <alexey.sachkov@intel.com>