Temporarily disable "native" GPU launches of atomic max/min
This is required for merging the code for P2 CMPLRLLVM-21741
before the GPU driver that supports min/max natively is installed
in alloy. Once the latter is done, the change should be reverted.

Before this disabling, manual tests of the whole feature were run
by the change owner (DPC++ engineering build & correct GPU driver
were used). This was tested on Gen9 and Arctic Sound GPUs.

Signed-off-by: Artem Gindinson <artem.gindinson@intel.com>
Change-Id: I694d4d249d0c09eab9bcd10508349351dd3e3743
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/306644
Tested-by: ICL Jenkins <icljenkins@eclists.intel.com>
Reviewed-by: Chupin, Pavel V <pavel.v.chupin@intel.com>
Reviewed-by: yaman, wang <yaman.wang@intel.com>